Fix types of inputs

+++ metrics.py
@@ -23,8 +23,8 @@ def get_model_performance(model, loader):
         # print(f"{batch_idx + 1} / {len(loader)}")
         # if counter > 10:
         #  break
-        inputs = inputs.to(device, dtype=torch.float)
-        labels = labels.to(device, dtype=torch.float)
+        inputs = inputs.to(device, dtype=torch.float64)
+        labels = labels.to(device)
         preds = model(inputs)["out"]  # make prediction
         labels_masked_land = torch.where(labels == 254, np.nan, labels)
         labels_masked_all = torch.where(
